requirements.txt
PyPDF2==3.0.1
langchain==0.1.0
langchain_google_genai
google-generativeai
langchain_community
python-dotenv
streamlit
faiss-cpu
nomic
gpt4all